requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.1.1
httpx==0.27.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
httpx==0.27.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
stripe==13.0.0
python-multipart==0.0.9
openai>=1.37.0

